IGNITE-11465 Multiple client leave/join events may wipe affinity assignment history and cause transactions fail
